requirements.txt
# Core dependencies
python-dotenv==1.2.1
langchain-core==1.1.0
langgraph==1.0.3

# Machine learning runtime
torch==2.3.1
numpy>=1.24.0
sentence-transformers>=2.2.0

# OpenAI Integration
langchain-openai>=0.2.0
openai>=1.0.0

# Ollama Integration
langchain-ollama==1.0.0

# Pinecone for vector search
langchain-pinecone>=0.1.1
pinecone>=5.0.1

# Flask API
Flask>=3.0.0
Flask-CORS>=4.0.0
